+++ R/qsc.R
@@ -23,7 +23,7 @@ vb <-     rbind(cbind(rbind(xy, -radius), rbind(xy, radius)),
               c(3, 7, 8, 4),
               c(5, 6, 7, 8))
   structure(list(vb = vb, ib = ib,
-                 primitivetype = "quad",
-                 material = list()), class = c("mesh3d", "shape3d"))
+
+                 material = list(), meshColor = "faces"), class = c("mesh3d", "shape3d"))
 }