+++ R/triangmesh.R
@@ -22,7 +22,7 @@
 #' may be set explicitly with `texture_filename`.  Currently it's not possible to not use the `texture` object
 #' in-memory.
 #' @inheritParams quadmesh
-#' @return mesh3d (primitivetype triangle)
+#' @return mesh3d (primitive type triangle)
 #' @export
 #' @aliases dtriangmesh
 #' @examples
@@ -71,7 +71,7 @@ triangmesh.BasicRaster <- function (x, z = x, na.rm = FALSE, ..., texture = NULL
 
 
   if (is.null(z)) z <- 0 else z <- raster::values(x)
-  ob <- structure(list(it = ind1, vb = t(cbind(coords, z, 1)), primitivetype = "triangle", material = list()), class = c("mesh3d", "shape3d"))
+  ob <- structure(list(it = ind1, vb = t(cbind(coords, z, 1)),  material = list(), meshColor = "faces"), class = c("mesh3d", "shape3d"))